requirements.txt
langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.24
langchain-pinecone==0.2.8
langchain-google-genai==2.0.5
google-generativeai==0.8.5
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0
deep-translator==1.11.4 
transformers==4.46.2 
torch==2.5.1 
accelerate==1.1.1
langdetect==1.0.9

-e .